backend/requirements.txt
# Production Requirements
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
flask-limiter==3.5.0
tensorflow==2.15.0
tensorflow-hub==0.15.0
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
python-dotenv==1.0.0
soundfile==0.12.1
audioread==3.0.1
